Extract createList and define getIds function
