+++ src/reducers/createList.js
@@ -1,0 +1,18 @@
+const createList = (filter) => {
+  return (state = [], action) => {
+    if (action.filter !== filter) {
+      return state;
+    }
+    switch (action.type) {
+    case 'RECEIVE_TODOS':
+      return action.response.map(todo => todo.id);
+    default:
+      return state;
+    }
+  };
+};
+
+export default createList;
+
+export const getIds = (state) => state;
+
